Black tern 0.2
This version includes the velocity calculation of transitional water waves and a simple algorithm to divide the linear deep water waves from the linear transitional water waves. The user can also define where its data file is located, this adding the <path to file/name of file.txt> after running the program. As usual to create an executable you just need to run in your unix/osx/linux terminal gcc blacktern.c -o blacktern.

To run the program, go to the folder were blacktern is stored and type ./blacktern <path to the file with the wave data>.txt

The wave data must be given as explained in the "how to use" text file. See you in 0.3 really soon!.