BKT0_6b
Density function that varies with salinity, temperature and atmospheric pressure added. Values are sensible below the 3% of the total pressure in the column, still function added as could be useful later. Temperature does not appear to affect density greatly as waves will mix the column on their depth of interaction so this function its not included, as thermoclyne appear to be fixed at depths that are on the last (1/3)*500m of water column even in summer according to information.

Might just pass a check on data to see  temperature measurements of the water column.

Now the program needs a file called pdata.txt, on which the sea temperature, sea salinity and atmospheric pressure are stored. The file must simply have:

<Temperature celsius> <Salinity PSU/PPT> <Atmospheric pressure on atmospheres> 

File must be stored at the same direction as the program, there is an already loaded file in the package, so it just need its values replaced by the ones of the user.