So-far
The models implemented based on the calculation of Airy, Stokes to third order (with some reductions) and Robert E. Jensen can be seen in the image, solutions in the space of the Korteweg de Vries space are the next ones to be added. Solution for the 4th and 5th order following the solutions done by Fenton and Robert E Jensen have to be verified yet this for 4th and 5th wave orders. After Korteweg de Vries is added current interaction will be added to to follow with a possible mapping of the non-linear interaction using analysis.